backend/requirements.txt
Django==2.0.5
# slightly older version forced by dynamic-rest
djangorestframework==3.7.7
django-environ==0.4.4
prettyrepr==2018.03.08.3
psycopg2-binary==2.7.4
tqdm==4.19.6
uWSGI==2.0.17
raven==6.6.0
factory-boy==2.10.0
logging-tree==1.7
tabulate==0.8.2
djangorestframework-jwt==1.11.0
djoser==1.1.5
dynamic-rest==1.8.1
dry-rest-permissions==0.1.10
